requirements.txt
discord.py
python-dotenv
asyncio
PyNaCl
pytubefix
logging
cysystemd
